pyproject.toml
[project]
name = "slide-extractor"
version = "0.1.0"
description = "Extract presentation slides from lecture videos into PDFs"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.83.0",
    "imagehash>=4.3.2",
    "img2pdf>=0.6.3",
    "pillow>=12.1.1",
    "rich>=14.3.3",
    "yt-dlp>=2026.2.21",
]

[project.scripts]
slide-extractor = "slide_extractor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slide_extractor"]

[dependency-groups]
dev = []
